requirements.txt
asgiref==3.8.1
attrs==24.3.0
bleach==6.2.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
webencodings==0.5.1
whitenoise==6.8.2
